refactor(order-resolution): 行軍命令解決処理の主要部分を Adjudicator に分離
