+++ server/src/domain/phase.rs
@@ -1,3 +1,4 @@
+mod adjudicator;
 mod order_helper;
 mod order_resolution;
